Extract trino-filesystem module
