+++ lib/trino-filesystem/src/main/java/io/trino/filesystem/FileIterator.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.iceberg.io;
+package io.trino.filesystem;
 
 import java.io.IOException;
 import java.util.NoSuchElementException;
