+++ lib/trino-filesystem/src/main/java/io/trino/filesystem/TrinoInput.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.iceberg.io;
+package io.trino.filesystem;
 
 import io.airlift.slice.Slice;
 import io.airlift.slice.Slices;
