+++ lib/trino-filesystem/src/main/java/io/trino/filesystem/fileio/ForwardingFileIo.java
@@ -11,9 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.iceberg.io.fileio;
+package io.trino.filesystem.fileio;
 
-import io.trino.plugin.iceberg.io.TrinoFileSystem;
+import io.trino.filesystem.TrinoFileSystem;
 import org.apache.iceberg.io.FileIO;
 import org.apache.iceberg.io.InputFile;
 import org.apache.iceberg.io.OutputFile;
